+++ app/views/user_orders/show.html.erb
@@ -22,19 +22,21 @@
 <center>
   <table>
     <tr>
-      <th>Item</th>
-      <th>Sold By</th>
-      <th>Price</th>
+      <th>Item Name</th>
+      <th>Description</th>
+      <th>Image</th>
       <th>Quantity</th>
+      <th>Price</th>
       <th>Subtotal</th>
     </tr>
   <% @order.item_orders.each do |item_order|%>
     <tr>
     <section id = "item-<%=item_order.item_id%>">
         <td><p><%=link_to item_order.item.name, "/items/#{item_order.item_id}"%></p></td>
-        <td><p><%= link_to item_order.item.merchant.name, "/merchants/#{item_order.item.merchant.id}"%></p></td>
-        <td><p><%= number_to_currency(item_order.price)%></p></td>
+        <td><p><%= item_order.item.description %></p></td>
+        <td><p><img src=<%= item_order.item.image %> /></p></td>
         <td><p><%= item_order.quantity%></p></td>
+        <td><p><%= number_to_currency(item_order.price)%></p></td>
         <td><p><%= number_to_currency(item_order.subtotal)%></p></td>
       </section>
     </tr>
