+++ spec/features/cart/checkout_spec.rb
@@ -57,7 +57,7 @@ RSpec.describe 'Cart show' do
 
       click_on "Checkout"
 
-      expect(current_path).to eq("/orders/new")
+      expect(current_path).to eq("/profile/orders/new")
     end
   end
