--- spec/features/orders/creation_spec.rb
@@ -1,80 +1,0 @@
-# When I fill out all information on the new order page
-# And click on 'Create Order'
-# An order is created and saved in the database
-# And I am redirected to that order's show page with the following information:
-#
-# - Details of the order:
-
-# - the date when the order was created
-RSpec.describe("Order Creation") do
-  describe "When I check out from my cart" do
-    before(:each) do
-      user = User.create(name: "David", address: "123 Test St", city: "Denver", state: "CO", zip: "80204", email: "123@example.com", password: "password", role: 1)
-      allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(user)
-
-      @mike = Merchant.create(name: "Mike's Print Shop", address: '123 Paper Rd.', city: 'Denver', state: 'CO', zip: 80203)
-      @meg = Merchant.create(name: "Meg's Bike Shop", address: '123 Bike Rd.', city: 'Denver', state: 'CO', zip: 80203)
-      @tire = @meg.items.create(name: "Gatorskins", description: "They'll never pop!", price: 100, image: "https://www.rei.com/media/4e1f5b05-27ef-4267-bb9a-14e35935f218?size=784x588", inventory: 12)
-      @paper = @mike.items.create(name: "Lined Paper", description: "Great for writing on!", price: 20, image: "https://cdn.vertex42.com/WordTemplates/images/printable-lined-paper-wide-ruled.png", inventory: 3)
-      @pencil = @mike.items.create(name: "Yellow Pencil", description: "You can write on paper with it!", price: 2, image: "https://images-na.ssl-images-amazon.com/images/I/31BlVr01izL._SX425_.jpg", inventory: 100)
-
-      visit "/items/#{@paper.id}"
-      click_on "Add To Cart"
-      visit "/items/#{@paper.id}"
-      click_on "Add To Cart"
-      visit "/items/#{@tire.id}"
-      click_on "Add To Cart"
-      visit "/items/#{@pencil.id}"
-      click_on "Add To Cart"
-
-      visit "/cart"
-      click_on "Checkout"
-    end
-
-    it 'I can create a new order' do
-      user = User.create(name: "David", address: "123 Test St", city: "Denver", state: "CO", zip: "80204", email: "123@example.com", password: "password", role: 1)
-      allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(user)
-
-
-      visit '/cart'
-      click_on 'Checkout'
-
-      fill_in :name, with: user.name
-      fill_in :address, with: user.address
-      fill_in :city, with: user.city
-      fill_in :state, with: user.state
-      fill_in :zip, with: user.zip
-
-      click_button "Create Order"
-
-      new_order = Order.last
-
-      expect(current_path).to eq("/profile/orders")
-
-      click_on "#{new_order.id}"
-      expect(current_path).to eql("/profile/orders/#{new_order.id}")
-
-    end
-
-    it 'i cant create order if info not filled out' do
-      name = ""
-      address = "123 Sesame St."
-      city = "NYC"
-      state = "New York"
-      zip = 10001
-
-      fill_in :name, with: name
-      fill_in :address, with: address
-      fill_in :city, with: city
-      fill_in :state, with: state
-      fill_in :zip, with: zip
-
-      click_button "Create Order"
-
-      expect(page).to have_content("Please complete address form to create an order.")
-      expect(page).to have_button("Create Order")
-    end
-
-
-  end
-end
