+++ spec/features/user_orders/create_spec.rb
@@ -1,9 +1,11 @@
-RSpec.describe("New Order Page") do
+require 'rails_helper'
+
+RSpec.describe("Order Creation") do
   describe "When I check out from my cart" do
     before(:each) do
       user = User.create(name: "David", address: "123 Test St", city: "Denver", state: "CO", zip: "80204", email: "123@example.com", password: "password", role: 1)
       allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(user)
-      
+
       @mike = Merchant.create(name: "Mike's Print Shop", address: '123 Paper Rd.', city: 'Denver', state: 'CO', zip: 80203)
       @meg = Merchant.create(name: "Meg's Bike Shop", address: '123 Bike Rd.', city: 'Denver', state: 'CO', zip: 80203)
       @tire = @meg.items.create(name: "Gatorskins", description: "They'll never pop!", price: 100, image: "https://www.rei.com/media/4e1f5b05-27ef-4267-bb9a-14e35935f218?size=784x588", inventory: 12)
@@ -18,50 +20,55 @@ RSpec.describe("New Order Page") do
       click_on "Add To Cart"
       visit "/items/#{@pencil.id}"
       click_on "Add To Cart"
-    end
-    it "I see all the information about my current cart" do
 
       visit "/cart"
-
       click_on "Checkout"
-
-      within "#order-item-#{@tire.id}" do
-        expect(page).to have_link(@tire.name)
-        expect(page).to have_link("#{@tire.merchant.name}")
-        expect(page).to have_content("$#{@tire.price}")
-        expect(page).to have_content("1")
-        expect(page).to have_content("$100")
-      end
-
-      within "#order-item-#{@paper.id}" do
-        expect(page).to have_link(@paper.name)
-        expect(page).to have_link("#{@paper.merchant.name}")
-        expect(page).to have_content("$#{@paper.price}")
-        expect(page).to have_content("2")
-        expect(page).to have_content("$40")
-      end
-
-      within "#order-item-#{@pencil.id}" do
-        expect(page).to have_link(@pencil.name)
-        expect(page).to have_link("#{@pencil.merchant.name}")
-        expect(page).to have_content("$#{@pencil.price}")
-        expect(page).to have_content("1")
-        expect(page).to have_content("$2")
-      end
-
-      expect(page).to have_content("Total: $142")
     end
 
-    it "I see a form where I can enter my shipping info" do
-      visit "/cart"
-      click_on "Checkout"
+    it 'I can create a new order' do
+      user = User.create(name: "David", address: "123 Test St", city: "Denver", state: "CO", zip: "80204", email: "123@example.com", password: "password", role: 1)
+      allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(user)
 
-      expect(page).to have_field(:name)
-      expect(page).to have_field(:address)
-      expect(page).to have_field(:city)
-      expect(page).to have_field(:state)
-      expect(page).to have_field(:zip)
+
+      visit '/cart'
+      click_on 'Checkout'
+
+      fill_in :name, with: user.name
+      fill_in :address, with: user.address
+      fill_in :city, with: user.city
+      fill_in :state, with: user.state
+      fill_in :zip, with: user.zip
+
+      click_button "Create Order"
+
+      new_order = Order.last
+
+      expect(current_path).to eq("/profile/orders")
+
+      click_on "#{new_order.id}"
+      expect(current_path).to eql("/profile/orders/#{new_order.id}")
+
+    end
+
+    it 'i cant create order if info not filled out' do
+      name = ""
+      address = "123 Sesame St."
+      city = "NYC"
+      state = "New York"
+      zip = 10001
+
+      fill_in :name, with: name
+      fill_in :address, with: address
+      fill_in :city, with: city
+      fill_in :state, with: state
+      fill_in :zip, with: zip
+
+      click_button "Create Order"
+
+      expect(page).to have_content("Please complete address form to create an order.")
       expect(page).to have_button("Create Order")
     end
+
+
   end
 end
